Fix all data flow issues for new dataset for arcs track (#45)
* Add generalized UI for updating background data sources

* Fix all data flow issues for new dataset for arcs track
