+++ src/components/enhancer-regions/EnhancerRegions.tsx
@@ -348,6 +348,7 @@ const EnhancerRegion = React.memo((_props) => {
       selectedSamples,
       numSamples,
       stratification,
+      enhancerTrackConfig,
     ]
   );
